pseudo_client.c: retry after a couple of milliseconds
Race conditions exist when the server shutdown takes long enough for
three attempts to access the server to fail. Solution: Add a slight
delay to the retry. Delay is variable (using getpid()%5). (Not "random"
because I have no evidence that the process the client is running in
will have seeded RNG, and I don't want to seed it and possibly screw
them up).