Server launch reworking.
This is the big overhaul to have the server provide meaningful exit status
to clients.

In the process, I discovered that the server was running with signals blocked
if launched by a client, which is not a good thing, and prevented this from
working as intended.

Still looking to see why more than one server spawn seems to happen.