make bash stop bullying me
So we had this really strange problem where, sometimes but not always,
pseudo would have strange problems on startup, where the pseudo server
would end up running under pseudo. And this produced the most fascinating
thing, which was:
	unsetenv("LD_PRELOAD");
	assert(getenv("LD_PRELOAD") == NULL);
	for (int i = 0; environ[i]; ++i) {
		assert(strncmp(environ[i], "LD_PRELOAD=", 11));
	}

(pseudocode untested)

This would crash on the environ search. Because getenv() was not searching
environ.

WHAT.

So it turns out, *bash overrides getenv, setenv, and so on*. Under those
names. Hiding the glibc ones. And this creates horrible problems if you
assumed that your code could call those functions and expect them to work.

So as a workaround, pseudo now uses dlsym to find getenv, etc., from
glibc, and invokes those directly if possible. Also the client now uses
unwrapped fork/exec for spawning the server, which cleans up the
behavior of that code quite a bit.