Don't try to record 0-length posix_acl_default xattrs
Based on a submission from Anton Gerasimov <anton@advancedtelematic.com>

On some systems, with some kernel configs, "cp -a" apparently tries to
set an empty ACL list, with a valid header but no contents, which causes
strange and mysterious behavior later if we actually create such an entry.
So filter that out, also sanity-check a couple of other things.

Signed-off-by: Seebs <seebs@seebs.net>