Prevent bash from segfaulting when unloading pseudo
bash's extremely fancy internal awareness of how the environment looks
means that, if you directly call the underlying libc "unsetenv" on
a variable, bash can end up trying to access a null pointer. Fixing
this generically is actually rather hard; you can't really avoid
writing to environ on fork() or popen(), even if you change all
execv*() functions to use the execv*e() variants. So for now, instead
of unsetting the variable, set it to an empty string.

Thanks to Saur in IRC for spotting this and helping debug it.

Signed-off-by: Seebs <seebs@seebs.net>