handle O_TMPFILE and linkat()
This is a rework which replaces a previous patch. In this version,
files created with O_TMPFILE don't get recorded in the database
at all, but if we get a link request for /proc/self/fd/N, and the
corresponding file is not in the database, we send a CREAT request
for it instead of a LINK, and that appears to work with a MUCH
reduced chance of database leakage.

Also the O_TMPFILE won't be creating bogus database entries
anymore.

Signed-off-by: Seebs <seebs@seebs.net>

linkat fix