Epoll: use the correct client
So in the non-listen-fd case, apparently I broke the code to use
the index of the returned event, rather than the corresponding ID
value registered with epoll, as the index into the clients array.
This, of course, was nonsense, and explains why we were seeing a
ton of attempts to modify client 0.

This also reverts the previous, incorrect, attempt to address all
those messages.

Signed-off-by: seebs <seebs@seebs.net>