Don't delete things and not create them.
When OP_CREAT comes in, if there's an existing entry with the
same inode, we nuke that entry because CREAT implies a new file
got that inode, and if there's an existing entry with the same
path, we don't create a link, because there's already something
there.

So if there's a single entry with the same path and inode, it
gets deleted and not recreated. Oops.

Now distinguishing between a single exact match of both path
and inode, and distinct matches by path and by inode.

This addresses a use case where parallel creations could
in theory cause pseudo to mistakenly drop database entries,
potentially. (I could only trigger it by doing this to files
which had existing entries, and then deleting the files outside
of pseudo, though.)

Signed-off-by: seebs <seebs@seebs.net>