Improve path handling by not dereferencing files
The path handling was stripping /., or moving back up a level for
/.., even when the current thing was a plain file, resulting in the
surprising behavior that "a/." was a valid name for a file called a,
which canonicalized without the extra directory. This breaks some
stuff occasionally. (Related to, but not the same as, the previous
issues with the use of trailing slashes in names to match only
directories.)

This is not yet well-tested.

Signed-off-by: Seebs <seebs@seebs.net>