attempt to handle large inode values
So ino_t is uint64_t, sqlite field types are magical and erratic and
if a value doesn't fit inside signed int64, it gets converted to
float64. of course. This means that the top half of the inode range
produces invalid-ish values, because a 64-bit double is typically
53 bits of mantissa, so a value around 2^64 will convert to a float64
value which is going to compare equal to at least 2^10 other integer
values if they were also converted to float64. Which makes weird
problems.