use *correct* flags for open/openat, also apply them for related stat
When statting a file that we may or may not be opening with O_NOFOLLOW,
we should use lstat (or AT_SYMLINK_NOFOLLOW) to try to get information
about the right file.

Also when we want to check whether a bit is set, we should use & rather
than |. I am an experienced programmer and know the difference between
those.