makewrappers: support architecture-overrides in wrapper modifiers
Pseudo allows wrappers to define special comments in the wrapper lists to
pass extra arguments such as version=GLIBC_2.3 to control which symbol
version to search for.  However, these arguments can be architecture-specific.

When parsing the arguments, check for flags that end in the architecture name
(as returned by platform.machine()) and use those values instead.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>