pseudo_client: Ensure renames update open fd file paths
There is an issue in pseudo where if you open a file, rename the file,
then call fstat on the open fd, pseudo would thrown an abort. This is
because it needs to track the open fd mappings to files and it doesn't
update in the case of a rename.

Add code in pseudo to update the fd mappings in the case of a rename
call. Also add a test case.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>