ports/rename/renameat: Avoid race when renaming files
We're seeing failures where a file is renamed under pseudo but an
access appears to be made to the old filename before the OP_RENAME has
hit the database but after the real_rename has applied in the kernel.

This is effectively the MAY_UNLINK problem for the original filename. There
were protections for the newpath but not the oldpath.

To try and avoid these aborts(), mark the original path as MAY_UNLINK however
we need to be careful not to trigger the DID_UNLINK but instead update
the deleting entry in the database as the rename completes. To do this,
we no clear the deleting flag during the database rename operation in SQL.

Also, we have to stop removing the by_ino matches where by_ino.deleting is
set, else we may lose the renamed file's attributes that way.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>