pseudo_client: Make msg static in pseudo_op_client
The address of msg is returned by the function in the OP_CHROOT case.
Whilst the way it is used means it is just a way of saying true/false,
the compiler doesn't know this and can warn and this isn't really
allowed.

The other pseudo funcitons in this area already use a static msg so
make this function match the others. There wouldn't be re-enterancy
in this context.

Need to be careful about clearing the data upon each function entry
and not just once at init.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>