pseudo_db: Flush DB if there is a shutdown request
We have some challenges in ensuring the pseudo database is flushed to disk
in things like docker containers since the processes can be killed with no
warning at container termination. Rightly or wrongly, we need to handle this
better.

There is no current way to flush the DB but there is a shutdoen request
mechanism. Whilst we can't force a shutdown from bitbake due to the multiple
connected clients and the existing 3s shutdown isn't enough, we can take the
hint to flush the DB at this time. Further writes are unlikely.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>