pseudo_util: Improve handling of crazy length paths
Crazy shell code (e.g. libtool) can pass in a command pipeline as a
path which exceeds the max path length the system can support (6000+
chars). This will fail in libc or the syscall but if we don't do
something here, we'd segfault before it can do that. Leave path
unchanged and let libc deal with it.

This was observed with segfaults in libfm:do_install after the libtool
upgrade. It does depend on the length of the local build path too.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>