Bugfix for Linux open(O_CREAT|O_EXCL)
In Linux, when calling open(path, O_CREAT|O_EXCL) it implies O_NOFOLLOW.
If the path exists and is a symlink, the link shouldn't be resolved but
rather just fail directly with EEXIST. Resolving the symlink causes
problems with e.g. extracting files with GNU tar where strange errors
ensue when it tries to overwrite an existing symlink.

Fix the bug by not resolving symlinks for open (and the related
functions) if O_EXCL is set.

For original patch from Simon Lindholm see:
See: https://bugzilla.yoctoproject.org/show_bug.cgi?id=15559

Original submission was modified to specifically check for BOTH
O_CREAT|O_EXCL, as this matches the behavior in the Linux kernel:

   fs/open.c: build_open_flags(...)

        if (flags & O_CREAT) {
                op->intent |= LOOKUP_CREATE;
                if (flags & O_EXCL) {
                        op->intent |= LOOKUP_EXCL;
                        flags |= O_NOFOLLOW;
                }
        }

If the kernel sees O_CREAT and O_EXCL, it assumes O_NOFOLLOW, we need to
duplicate the same behavior.  Note, this behavior is likely Linux specific
thus implemented in the ports/linu/wrapfuncs.in.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@kernel.crashing.org>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>